Added the Universal Control Center.
The UCC is a persistent tmux session running in workspace 0.
It is toggled by super + Enter, which was previously used to open a shell.
Furthermore, most i3 hotkeys were moved to sxhkd, except for some that
were removed and the distraction-free mode hotkey (super + shift + f).

This is version 1 of the UCC, more updates are to follow.